+++ Jiten.Api/Dtos/Requests/DeckDownloadRequest.cs
@@ -10,5 +10,5 @@ public class DeckDownloadRequest
     public bool ExcludeKana { get; set; }
     public bool ExcludeKnownWords { get; set; }
     public bool ExcludeExampleSentences { get; set; }
-    public List<int>? KnownWordIds { get; set; }
+    public List<long>? KnownWordIds { get; set; }
 }
